+++ example/android/app/src/main/kotlin/com/example/quick_toast_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.quick_toast_example
+package com.example.mini_toast_example
 
 import io.flutter.embedding.android.FlutterActivity